requirements.txt
pyspellchecker==0.7.2
unidecode==1.3.6
